requirements.txt
paho-mqtt==1.6.1
pymodbus==2.5.3
pyModbusTCP==0.1.10
